P-34 Make config data as user secret
