+++ backend/services/GatherService/Controllers/ChannelsController.cs
@@ -1,8 +1,7 @@
 ﻿using Asp.Versioning;
 using Gather.Dtos;
 using Gather.Models;
-using Gather.Services.ChannelsService;
-using Gather.Services.InfoService;
+using Gather.Services.Channels;
 using Microsoft.AspNetCore.Mvc;
 using TL;
 
@@ -173,4 +172,66 @@ public class ChannelsController(IChannelsService channelsService) : ControllerBa
         }
     }
 
+
+
+    [HttpGet]
+    [Route("{channelId}/posts/{postId}/comments")]
+    [MapToApiVersion(1.0)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    public async Task<ActionResult<List<CommentDto>>> GetAllPostComments(long channelId, long postId, int offset = 0, int count = 20)
+    {
+        var response = await _channelsService.GetComments(channelId, postId, offset, count);
+
+        if (response.Message == "Channel not found")
+        {
+            return NotFound(response);
+        }
+
+        if (!response.Success)
+        {
+            return StatusCode(StatusCodes.Status500InternalServerError, response);
+        }
+
+        return Ok(response);
+    }
+
+    [HttpGet()]
+    [Route("{channelId}/posts/{postId}/update_comments")]
+    public async Task UpdatePostComments(long channelId, long postId)
+    {
+        if (HttpContext.WebSockets.IsWebSocketRequest)
+        {
+            using var webSocket = await HttpContext.WebSockets.AcceptWebSocketAsync();
+            await _channelsService.UpdatePostComments(channelId, postId, webSocket);
+        }
+        else
+        {
+            HttpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
+        }
+    }
+
+    [HttpGet]
+    [Route("{channelId}/posts/{postId}/comments_count")]
+    [MapToApiVersion(1.0)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    public async Task<ActionResult<long>> GetPostCommentsCount(long channelId, long postId)
+    {
+        var response = await _channelsService.GetCommentsCount(channelId, postId);
+
+        if (response.Message == "Channel not found")
+        {
+            return NotFound(response);
+        }
+
+        if (!response.Success)
+        {
+            return StatusCode(StatusCodes.Status500InternalServerError, response);
+        }
+
+        return Ok(response);
+    }
 }
