+++ backend/services/GatherService/Services/Search/ISearchService.cs
@@ -1,0 +1,9 @@
+﻿using Gather.Models;
+
+namespace Gather.Services.Search;
+
+public interface ISearchService
+{
+
+    Task<ServiceResponse<object>> Search(SearchQuery query);
+}
