--- backend/services/GatherService/Services/SearchService/ISearchService.cs
@@ -1,10 +1,0 @@
-﻿using Gather.Models;
-
-namespace Gather.Services.SearchService
-{
-    public interface ISearchService
-    {
-
-        Task<ServiceResponse<object>> Search(SearchQuery query);
-    }
-}
