+++ backend/services/GatherService/Utils/Config/ConfigUtils.cs
@@ -1,0 +1,19 @@
+﻿namespace Gather.Utils.ConfigService;
+
+public class ConfigUtils(string apiId, string apiHash, string phoneNumber) : IConfigUtils
+{
+    public Func<string, string> ConfigMethod = what =>
+    {
+        if (what == "api_id") return apiId;
+        if (what == "api_hash") return apiHash;
+        if (what == "phone_number") return phoneNumber;
+        if (what == "verification_code") return null; // let WTelegramClient ask the user with a console prompt
+        //if (what == "password") return "password";     // if user has enabled 2FA
+        return null;
+    };
+
+    public Func<string, string> Config()
+    {
+        return ConfigMethod;
+    }
+}
