+++ frontend/src/app/styles/index.module.css
@@ -22,5 +22,4 @@
     flex: 6 6 85vw;
     box-shadow: var(--shadow);
     border-radius: .3rem;
-    padding: .5rem;
 }
