--- frontend/src/entities/Props/CommentsLoadingWidgetProps.ts
@@ -1,6 +1,0 @@
-export interface CommentsLoadingWidgetProps {
-    channelId: number | undefined;
-    postId: number;
-    invalidateCashe: () => void;
-    setLoadingError: (description: string) => void;
-}
