+++ frontend/src/entities/Props/CommentsUpdatingWidgetProps.ts
@@ -1,0 +1,6 @@
+export interface CommentsUpdatingWidgetProps {
+    channelId: number | undefined;
+    postId: number;
+    invalidateCashe: () => void;
+    setUpdatingError: (description: string) => void;
+}
