+++ frontend/src/entities/Props/index.ts
@@ -2,6 +2,6 @@ export type { BadgeProps } from './BadgeProps';
 export type { ChannelMainInfoProps } from './ChannelMainInfoProps';
 export type { ChannelProps } from './ChannelProps';
 export type { PostsLoadingWidgetProps } from './PostsLoadingWidgetProps';
-export type { CommentsLoadingWidgetProps } from './CommentsLoadingWidgetProps';
+export type { CommentsUpdatingWidgetProps } from './CommentsUpdatingWidgetProps';
 export type { PostsProps } from './PostsProps';
 export type { UsersProps } from './UsersProps';
