+++ frontend/src/entities/comments/api/get-comments.ts
@@ -19,7 +19,7 @@ export const getComments = async (
 
     const result = await apiClient
         .get<ServiceResponse<CommentDto[]>>
-        (`info/channels/${channelId}/posts/${postId}/comments?offset=${offset}&limit=${limit}`);
+        (`channels/${channelId}/posts/${postId}/comments?offset=${offset}&limit=${limit}`);
 
     return ({
         comments: result.data.map((comment: CommentDto) => mapComment(comment))
@@ -31,7 +31,7 @@ export const getComment = async (channelId: string | undefined, postId: string |
         return Promise.resolve({ comment: null });
     }
 
-    const result = await apiClient.get<ServiceResponse<CommentDto>>(`info/channels/${channelId}/posts/${postId}/comments/${commentId}`);
+    const result = await apiClient.get<ServiceResponse<CommentDto>>(`channels/${channelId}/posts/${postId}/comments/${commentId}`);
 
     return ({
         comment: mapComment(result.data)
@@ -43,7 +43,7 @@ export const getPostCommentsCount = async (channelId: string | undefined, postId
         return Promise.resolve(0);
     }
 
-    const result = await apiClient.get<ServiceResponse<number>>(`info/channels/${channelId}/posts/${postId}/comments_count`);
+    const result = await apiClient.get<ServiceResponse<number>>(`channels/${channelId}/posts/${postId}/comments_count`);
 
     return (
         result.data as number
