--- frontend/src/shared/components/CommentsLoadingWidget/styles.module.css
@@ -1,20 +1,0 @@
-.block {
-    display: flex;
-    flex-direction: column;
-    row-gap: 1rem;
-    font-family: 'Roboto', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-    border-radius: 1rem;
-    box-shadow: rgba(0, 0, 0, 0.05) 0px 6px 24px 0px, rgba(0, 0, 0, 0.08) 0px 0px 0px 1px;
-    padding: 1rem
-}
-
-.block__text {
-    font-size: 1rem;
-}
-
-.error {
-    font-family: "Roboto", "Helvetica Neue", Helvetica, Arial, sans-serif;
-    font-size: medium;
-    font-weight: 500;
-    color: rgb(52, 71, 103);
-}
