+++ frontend/src/shared/components/CommentsUpdatingWidget/index.ts
@@ -1,0 +1,1 @@
+export { CommentsUpdatingWidget } from './CommentsUpdatingWidget';
